Check if a custom size for the primary key field is set
The hardcoded value 255 causes problems if we use a database collation that does
not support indices of 255, for instance 'utf8mb4_unicode_ci' (max length 191). This
will cause sails to not load the ORM as the database will throw an error like:

"Error: ER_TOO_LONG_KEY: Specified key was too long; max key length is X bytes"
In the case of mb4 the max length is 767 bytes (191*4).