Removing auto setting JSON to LONGTEXT
If a user tells the system they want to use columnType JSON, we should allow them to do so. By default, if no columnType is specified, then we default it to LONGTEXT. If a user attempts to use columntype JSON and their system does not support it, that is on them to fix the issue. Other wise, they can always not put a columnType and it will maintain backwards compatibility.